// src/api/models/collaboratorRole.ts
/**
 * Generated by orval v8.0.3 🍺
 * Do not edit manually.
 * EdCraft Backend API
 * API for EdCraft Backend
 * OpenAPI spec version: 0.1.0
 */

/**
 * Roles for resource collaborators.
 */
export type CollaboratorRole = typeof CollaboratorRole[keyof typeof CollaboratorRole];


export const CollaboratorRole = {
  owner: 'owner',
  editor: 'editor',
  viewer: 'viewer',
} as const;