// src/api/models/resourceVisibility.ts
/**
 * Generated by orval v8.0.3 🍺
 * Do not edit manually.
 * EdCraft Backend API
 * API for EdCraft Backend
 * OpenAPI spec version: 0.1.0
 */

/**
 * Visibility settings for resources (assessments, etc.).
 */
export type ResourceVisibility = typeof ResourceVisibility[keyof typeof ResourceVisibility];


export const ResourceVisibility = {
  private: 'private',
  public: 'public',
} as const;